NO-JIRA: C++ enable warnings as errors for examples and tests.
Fixed warnings for gcc++ and clang++, 03 and 11, including some real bugs.
Removed some redundant FIXME comments.

+++ examples/cpp/mt/broker.cpp
@@ -247,7 +247,7 @@ class broker {
     }
 
     void run() {
-        for(int i = 0; i < std::thread::hardware_concurrency(); ++i)
+        for(size_t i = 0; i < std::thread::hardware_concurrency(); ++i)
             std::thread(&proton::controller::run, controller_.get()).detach();
         controller_->wait();
     }
